tools/uniprot-retriever/requirements.txt
requests>=2.31.0
biopython>=1.81
pandas>=2.0.0
